+++ core/model/auth.go
@@ -67,10 +67,10 @@ type LoginSession struct {
 
 	Device Device
 
-	IPAddress    string
-	AccessToken  string
-	RefreshToken string
-	Params       map[string]interface{} //authType-specific set of parameters passed back to client
+	IPAddress     string
+	AccessToken   string
+	RefreshTokens []string
+	Params        map[string]interface{} //authType-specific set of parameters passed back to client
 
 	Expires      time.Time
 	ForceExpires *time.Time
@@ -83,6 +83,15 @@ type LoginSession struct {
 func (ls LoginSession) IsExpired() bool {
 	now := time.Now()
 	return ls.Expires.Before(now) || (ls.ForceExpires != nil && ls.ForceExpires.Before(now))
+}
+
+//CurrentRefreshToken returns the current refresh token (last element of RefreshTokens)
+func (ls LoginSession) CurrentRefreshToken() string {
+	numTokens := len(ls.RefreshTokens)
+	if numTokens <= 0 {
+		return ""
+	}
+	return ls.RefreshTokens[numTokens-1]
 }
 
 //APIKey represents an API key entity